+++ src/models/index.js
@@ -1,2 +1,2 @@
-export { default as Post } from './post'
+export { default as Post } from './post-model'
 export { default as User } from './user'
